add release notes
